+++ libminifi/include/Port.h
@@ -1,4 +1,5 @@
 /**
+ *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
  * this work for additional information regarding copyright ownership.
@@ -14,25 +15,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+#pragma once
 
-#include "Funnel.h"
-#include "core/ProcessSession.h"
+#include "ForwardingNode.h"
 
 namespace org::apache::nifi::minifi {
 
-const core::Relationship Funnel::Success("success", "FlowFiles are routed to success relationship");
+class Funnel final : public ForwardingNode {
+ public:
+  Funnel(std::string name, const utils::Identifier& uuid) : ForwardingNode(std::move(name), uuid, core::logging::LoggerFactory<Funnel>::getLogger()) {}
+  explicit Funnel(std::string name) : ForwardingNode(std::move(name), core::logging::LoggerFactory<Funnel>::getLogger()) {}
 
-void Funnel::initialize() {
-  setSupportedRelationships(relationships());
-}
-
-void Funnel::onTrigger(const std::shared_ptr<core::ProcessContext>& /*context*/, const std::shared_ptr<core::ProcessSession>& session) {
-  logger_->log_trace("On trigger %s", getUUIDStr());
-  std::shared_ptr<core::FlowFile> flow_file = session->get();
-  if (!flow_file) {
-    return;
-  }
-  session->transfer(flow_file, Success);
-}
+  MINIFIAPI static constexpr core::annotation::Input InputRequirement = core::annotation::Input::INPUT_ALLOWED;
+  ADD_COMMON_VIRTUAL_FUNCTIONS_FOR_PROCESSORS
+};
 
 }  // namespace org::apache::nifi::minifi
